🔧 Updating Solana to match the Anchor 0.25.0
To use solana 1.14.7 (the version the anchor is using), I've had to use the local build of the anchor 0.25.0. The reason is that the published version (https://crates.io/crates/anchor-client/0.25.0/dependencies) uses the older version of solana, and there were unsolvable conflicts during the solana update. Once the anchor is published with the solana 1.14.7 or newer, publish this branch but resolve all todos in cargo.toml files and in the explorer crates.